Miminum set of changes that allow me to regenerate figures
I'm a little dissapointed to see that in the newly regenerated Fig 2 I
get an R^2 of .44 instead of the expected .49. Not sure whats going on
here because as far as I can tell this should be the code exaclty as
Monika had left it, with only superficial changes such as
wrapping things in a function and pointing to the correct filenams.

But its pretty close.
I plan to use this as a jumping off point to reimpement the minumum
changes I want to change how e do normalizaiton, interpolation and smoothing.