photobleaching correction by dividing off exponential fit
I've added the photobleachign correction code. I am photobleaching
following the logic outlined in Xiaowen Chen's Phys Rev E paper 2019.
This is slightly different then what Jeff did in the 3dBrain MATLAB code.
In that code he fit the exponential and then subtracted it off. Here I
divide it off, but rescale the trace to match the original.

Note: in its current state, the pipeline is not utilizing the
photocorrected traces, but now at least I have the machinery to create
it.