Fixed algebra error in photobleaching correction
Previously the photobleaching correction was failing by producing
traces that blew up towards the latter half of the trace.

The bug was an algebra error. To fit the exponent, at Francesco's
suggestion, I rescaled the origional traces so that it was on the range
of zero to one. This dramatically improves fitting performance and
efficiency. So then I had to rescale the fitting parameters back.

Previously I had been only scaling the a scaler in y=a exp(-bx)+ c.
But of course c had been scaled to. So now I also rescale c. And that
fixes everything.

ADDITONAL:
fixed a bug about using max instead of nanmax in the fitting routines.