Now using I_smooth_interp_crop_noncontig without runtime errors.
Switched most of the pipeline to use a noncontigouos, smooth, interpolated
activity trace that is not z-scored. Time points that are mostly NaNs
are excised from the time series.

I also went through and added code to raise runTimeErrors of some of the code
that hasn't been updated is called.